implement address management based on location data

+++ src/main/kotlin/model/LocationReq.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class LocationReq(
     @SerialName("horizontal_accuracy")
-    val horizontalAccuracy: Double,
+    val horizontalAccuracy: Int,
     @SerialName("is_precise_location_denied")
     val isPreciseLocationDenied: Boolean,
     @SerialName("lon")
